apps/agent_service/pyproject.toml
[project]
name = "evento-agent-svc"
version = "0.1.0"
description = "Multi-agent event content creation service using Google ADK"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.27.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.4.0",
    "firebase-admin>=6.5.0",
    "python-dotenv>=1.0.1",
    "websockets>=15.0.1",
    "google-cloud-firestore>=2.25.0",
    "google-cloud-storage>=3.9.0",
    "google-genai>=1.67.0",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "pytz>=2026.1.post1",
]

